remove card and token-specific structs
this attempts to covert the parser back into a lexer that outputs a vec of tokens. Being less rigid with this stream should set us up to provide more informative error messages later. Instead of discarding whole malformed cards it should be relatively cheap to determine malformed cards with concrete remediations in a second parser over this output stream