begin parser rewrite with nom
this begins re-implementing the parser with nom. I have more experience with nom, and at least up to the point of getting a parser basically running against a test suite I'm more confident that I can debug this.

Chumsky has a reputation for great error support, and for the smaller parsers passing around span information seemed like it would make things easier when it came time to connect the parser to the language server.

Nom has official documentation about this, and it's something to investigate after we get a parser running with a robust test suite